Add workaround for Acalog bug that encodes identifier
- There is a bug in the Acalog Publisher interface where the text of our
  common course identifier (ampersand) is html encoded before displaying
in the input field. If the course editor doesn't see this and saves,
then the encoded value becomes part of the value. Acalog says it is
working on a fix for that, but in the meantime, I'm adding one decode to
catch most cases. Unfortunately, the Acalog interface will continue to
commit html encoding inception on this value, so a course saved twice
won't be saved by this fix.